Address Windows warning with shift

+++ src/pos/ProofCore.hpp
@@ -506,7 +506,7 @@ public:
     // returns num/denom pair for expected pruned entries for t3
     std::pair<uint64_t, uint64_t> nd_expected_pruned_entries_for_t3()
     {
-        uint64_t k_entries = 1UL << static_cast<uint64_t>(params_.get_k());
+        uint64_t k_entries = 1ULL << params_.get_k();
         uint64_t numerator = FINAL_TABLE_FILTER * 4 * k_entries;
         uint64_t denominator = (1UL << 32);
         return std::make_pair(numerator, denominator);
